move auto download to type folder
