+++ modules/services/model/src/main/java/au/com/shiftyjelly/pocketcasts/models/type/AutoDownloadLimitSetting.kt
@@ -1,4 +1,4 @@
-package au.com.shiftyjelly.pocketcasts.preferences.model
+package au.com.shiftyjelly.pocketcasts.models.type
 
 import androidx.annotation.StringRes
 import au.com.shiftyjelly.pocketcasts.localization.R as LR
